# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DocPatient Nexus API",
    "version": "2.0.0",
...
    "contact": {
      "name": "Miyaad Shah Joy",
      "email": "docpatientnexus@example.com"
    }
  },
  "servers": [
    {
      "url": "https://docpatient-nexus.onrender.com",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
  "components": {
    "schemas": {
...
      "Medication": {
        "type": "object",
        "required": [
          "name",
          "dosage",
          "frequency",
          "duration"
        ],
        "properties": {
          "name": {
            "type": "string",
            "example": "Aspirin"
          },
          "dosage": {
            "type": "string",
            "example": "500mg"
          },
          "frequency": {
            "type": "string",
            "example": "Twice a day"
          },
          "duration": {
            "type": "number",
            "example": 7
          },
          "instruction": {
            "type": "string",
            "example": "Take with water, do not exceed the recommended dosage."
          }
        }
      },
      "Prescription": {
        "type": "object",
        "required": [
          "doctor",
          "patient",
          "appointment",
          "medications"
        ],
        "properties": {
          "doctor": {
            "type": "string",
            "example": "682787f1fea3f44089558cd6"
          },
          "patient": {
            "type": "string",
            "example": "682787f1fea3f44089558cd7"
          },
          "appointment": {
            "type": "string",
            "example": "682787f1fea3f44089558cd8"
          },
          "notes": {
            "type": "string",
            "example": "Take the medication after meals."
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "expired",
              "deleted"
            ],
            "default": "active"
          },
          "medications": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Medication"
            },
            "example": [
              {
                "name": "Aspirin",
                "dosage": "500mg",
                "frequency": "Twice a day",
                "duration": 7,
                "instruction": "Take with water, do not exceed the recommended dosage."
              },
              {
                "name": "Paracetamol",
                "dosage": "500mg",
                "frequency": "Once a day",
                "duration": 5,
                "instruction": "Take after meals."
              }
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-16T18:46:09.776Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2025-05-31T02:45:43.207Z"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
  "paths": {
...
    "/api/v2/doctors/appointments/{appointmentId}/prescription/": {
      "post": {
        "tags": [
          "Doctors"
        ],
        "summary": "Create a prescription for an appointment",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "description": "Allows a doctor to `create a prescription` for an appointment. A doctor must be `logged in` to use this route.",
        "operationId": "createPrescription",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "description": "ID of the appointment to create a prescription for.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "medications"
                ],
                "properties": {
                  "notes": {
                    "type": "string",
                    "example": "Patient advised to continue medications for 7 days and return for follow-up if symptoms persist."
                  },
                  "medications": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Medication"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "message": {
                      "type": "string",
                      "example": "Prescription created successfully."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Possibly due to invalid input.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Prescription already exists for this appointment."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Only `logged-in Doctors` can access this route. Log in with a valid `jwt` token.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You are not authorized to access this route. Please log in."
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Forbidden access. Only `logged-in Doctors` can access this route.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You do not have the permission to perform this action."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No appointment found with the provided ID.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "No appointment found with the provided ID."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Something went wrong while processing your request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v2/doctors/appointments/{appointmentId}/prescription": {
      "post": {
        "tags": [
          "Prescription"
        ],
        "summary": "Create a prescription for a patient.",
        "description": "**Allows a doctor to create a prescription for a patient after an appointment.**",
        "operationId": "createPrescription",
        "parameters": [
          {
            "name": "appointmentId",
            "in": "path",
            "required": true,
            "description": "ID of the appointment for which the prescription is created",
            "schema": {
              "type": "string",
              "example": "60c72b2f9b1e8b001c8e4d3a"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "medications"
                ],
                "properties": {
                  "medications": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Medication"
                    }
                  },
                  "notes": {
                    "type": "string",
                    "example": "Take the medication after meals."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Prescription created successfully. Returns the created prescription details.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "prescription": {
                          "$ref": "#/components/schemas/Prescription"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request. Possibly due to missing required fields or invalid data.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "Invalid input data. Please ensure all required fields are provided and correctly formatted."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized access. Doctor must be authenticated to create a prescription.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "You must be signed in as a doctor to create a prescription."
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No appointment found with the provided ID or appointment does not belong to the doctor.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "fail"
                    },
                    "message": {
                      "type": "string",
                      "example": "No appointment found with the provided ID or the appointment does not exist."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error. Something went wrong while processing your request.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "string",
                      "example": "error"
                    },
                    "message": {
                      "type": "string",
                      "example": "Something went wrong. Please try again later."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}